# executor: copy row data instead of refereeing chunk.Row in some window functions.