--- a/executor/aggfuncs/func_percent_rank.go
+++ b/executor/aggfuncs/func_percent_rank.go
@@ -31,31 +31,23 @@ func (pr *percentRank) AllocPartialResult() PartialResult {
 
 func (pr *percentRank) ResetPartialResult(partial PartialResult) {
 	p := (*partialResult4Rank)(partial)
-	p.curIdx = 0
-	p.lastRank = 0
-	p.rows = p.rows[:0]
+	p.reset()
 }
 
 func (pr *percentRank) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, partial PartialResult) error {
 	p := (*partialResult4Rank)(partial)
-	p.rows = append(p.rows, rowsInGroup...)
+	p.updatePartialResult(rowsInGroup, false, pr.compareRows)
 	return nil
 }
 
 func (pr *percentRank) AppendFinalResult2Chunk(sctx sessionctx.Context, partial PartialResult, chk *chunk.Chunk) error {
 	p := (*partialResult4Rank)(partial)
-	numRows := int64(len(p.rows))
-	p.curIdx++
-	if p.curIdx == 1 {
-		p.lastRank = 1
+	numRows := len(p.results)
+	if numRows == 1 {
 		chk.AppendFloat64(pr.ordinal, 0)
-		return nil
+	} else {
+		chk.AppendFloat64(pr.ordinal, float64(p.results[p.curIdx]-1)/float64(numRows-1))
 	}
-	if pr.compareRows(p.rows[p.curIdx-2], p.rows[p.curIdx-1]) == 0 {
-		chk.AppendFloat64(pr.ordinal, float64(p.lastRank-1)/float64(numRows-1))
-		return nil
-	}
-	p.lastRank = p.curIdx
-	chk.AppendFloat64(pr.ordinal, float64(p.lastRank-1)/float64(numRows-1))
+	p.curIdx++
 	return nil
 }