--- a/types/mydecimal.go
+++ b/types/mydecimal.go
@@ -250,6 +250,21 @@ func (d *MyDecimal) GetDigitsFrac() int8 {
 	return d.digitsFrac
 }
 
+// Copy copies a new *MyDecimal from itself.
+func (d *MyDecimal) Copy() *MyDecimal {
+	if d == nil {
+		return nil
+	}
+	dst := &MyDecimal{
+		digitsInt:  d.digitsInt,
+		digitsFrac: d.digitsFrac,
+		resultFrac: d.resultFrac,
+		negative:   d.negative,
+	}
+	copy(dst.wordBuf[:], d.wordBuf[:])
+	return dst
+}
+
 // String returns the decimal string representation rounded to resultFrac.
 func (d *MyDecimal) String() string {
 	tmp := *d