<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Fixes #11614, and fixes #11626 .
If a `chunk.Chunk` reuse into a window function executor, it may break the internal memory reference in function `RANK` or `PERCENT_RANK`.

1. Father calls `chk := NewChunk()`
2. Father calls `windowExecutor.Next(chk)`
3. The `windowExecutor` put column results into `chk` while refereeing the `chunk.Row` into [`PartialResult`](https://github.com/pingcap/tidb/blob/f98ef22/executor/aggfuncs/aggfuncs.go#L89) of  function `RANK` or `PERCENT_RANK`.
4. Father calls `windowExecutor.Next(chk)` again to collect next batch of result and reuses the same `chk`.
5. `chunk.Row` refereced by `PartialResult` is broken when `chk.Reset` is called.

- [x] TODO: fix the other window functions which may lead this problem.

### What is changed and how it works?

Change step 3 above, copy row data instead of refereeing chunk.Row in window functions which rely on data comparing.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Code changes

- None

Side effects

 - Possible performance regression

* note: no func `lag` because it cannot be finished with old code.
```
name                                                           old time/op    new time/op    delta
WindowRows/(func:row_number,_ndv:10,_rows:1000)-12               65.7µs ± 9%    68.5µs ± 8%       ~     (p=0.523 n=3+3)
WindowRows/(func:row_number,_ndv:1000,_rows:1000)-12              131µs ± 2%     161µs ± 8%    +22.55%  (p=0.041 n=3+3)
WindowRows/(func:row_number,_ndv:10,_rows:100000)-12             5.98ms ± 1%    5.97ms ± 1%       ~     (p=0.784 n=3+3)
WindowRows/(func:row_number,_ndv:1000,_rows:100000)-12           6.19ms ± 2%    6.07ms ± 0%       ~     (p=0.173 n=3+3)
WindowFunctions/(func:row_number,_ndv:1000,_rows:100000)-12      6.07ms ± 2%    6.06ms ± 3%       ~     (p=0.942 n=3+3)
WindowFunctions/(func:rank,_ndv:1000,_rows:100000)-12            6.79ms ± 0%    9.61ms ± 1%    +41.60%  (p=0.000 n=3+3)
WindowFunctions/(func:dense_rank,_ndv:1000,_rows:100000)-12      6.81ms ± 1%    9.70ms ± 1%    +42.59%  (p=0.000 n=3+3)
WindowFunctions/(func:cume_dist,_ndv:1000,_rows:100000)-12       6.82ms ± 3%   10.09ms ± 1%    +47.99%  (p=0.000 n=3+3)
WindowFunctions/(func:percent_rank,_ndv:1000,_rows:100000)-12    6.86ms ± 1%    9.62ms ± 0%    +40.25%  (p=0.000 n=3+3)
WindowFunctions/(func:ntile,_ndv:1000,_rows:100000)-12           6.31ms ± 1%    6.17ms ± 1%       ~     (p=0.050 n=3+3)
WindowFunctions/(func:lead,_ndv:1000,_rows:100000)-12            8.04ms ± 1%   12.30ms ± 0%    +52.91%  (p=0.000 n=3+3)
WindowFunctions/(func:first_value,_ndv:1000,_rows:100000)-12     6.28ms ± 1%    6.44ms ± 0%       ~     (p=0.063 n=3+3)
WindowFunctions/(func:last_value,_ndv:1000,_rows:100000)-12      6.36ms ± 0%    6.40ms ± 1%       ~     (p=0.415 n=3+3)
WindowFunctions/(func:nth_value,_ndv:1000,_rows:100000)-12       6.48ms ± 1%    6.47ms ± 1%       ~     (p=0.871 n=3+3)

name                                                           old alloc/op   new alloc/op   delta
WindowRows/(func:row_number,_ndv:10,_rows:1000)-12               44.7kB ± 0%    44.7kB ± 0%       ~     (zero variance)
WindowRows/(func:row_number,_ndv:1000,_rows:1000)-12             81.0kB ± 1%    80.5kB ± 1%       ~     (p=0.523 n=3+3)
WindowRows/(func:row_number,_ndv:10,_rows:100000)-12              263kB ± 0%     263kB ± 0%       ~     (p=0.443 n=3+3)
WindowRows/(func:row_number,_ndv:1000,_rows:100000)-12            301kB ± 0%     299kB ± 1%       ~     (p=0.423 n=3+3)
WindowFunctions/(func:row_number,_ndv:1000,_rows:100000)-12       301kB ± 0%     301kB ± 0%       ~     (zero variance)
WindowFunctions/(func:rank,_ndv:1000,_rows:100000)-12             307kB ± 0%    6917kB ± 0%  +2155.27%  (p=0.000 n=3+3)
WindowFunctions/(func:dense_rank,_ndv:1000,_rows:100000)-12       306kB ± 0%    6921kB ± 0%  +2160.67%  (p=0.000 n=3+3)
WindowFunctions/(func:cume_dist,_ndv:1000,_rows:100000)-12        306kB ± 0%    6923kB ± 0%  +2161.01%  (p=0.000 n=3+3)
WindowFunctions/(func:percent_rank,_ndv:1000,_rows:100000)-12     306kB ± 0%    6929kB ± 0%  +2162.99%  (p=0.000 n=3+3)
WindowFunctions/(func:ntile,_ndv:1000,_rows:100000)-12            307kB ± 0%     301kB ± 0%     -1.95%  (p=0.000 n=3+3)
WindowFunctions/(func:lead,_ndv:1000,_rows:100000)-12             306kB ± 0%    5984kB ± 0%  +1857.22%  (p=0.000 n=3+3)
WindowFunctions/(func:first_value,_ndv:1000,_rows:100000)-12      301kB ± 0%     301kB ± 0%       ~     (zero variance)
WindowFunctions/(func:last_value,_ndv:1000,_rows:100000)-12       301kB ± 0%     301kB ± 0%       ~     (zero variance)
WindowFunctions/(func:nth_value,_ndv:1000,_rows:100000)-12        301kB ± 0%     301kB ± 0%       ~     (zero variance)

name                                                           old allocs/op  new allocs/op  delta
WindowRows/(func:row_number,_ndv:10,_rows:1000)-12                 74.0 ± 0%      74.0 ± 0%       ~     (zero variance)
WindowRows/(func:row_number,_ndv:1000,_rows:1000)-12                698 ± 3%       690 ± 1%       ~     (p=0.523 n=3+3)
WindowRows/(func:row_number,_ndv:10,_rows:100000)-12              1.94k ± 0%     1.94k ± 0%       ~     (p=1.000 n=3+3)
WindowRows/(func:row_number,_ndv:1000,_rows:100000)-12            3.00k ± 0%     3.00k ± 0%       ~     (p=0.423 n=3+3)
WindowFunctions/(func:row_number,_ndv:1000,_rows:100000)-12       3.00k ± 0%     3.00k ± 0%       ~     (zero variance)
WindowFunctions/(func:rank,_ndv:1000,_rows:100000)-12             3.01k ± 0%    23.81k ± 0%   +691.91%  (p=0.000 n=3+3)
WindowFunctions/(func:dense_rank,_ndv:1000,_rows:100000)-12       3.01k ± 0%    23.83k ± 0%   +692.33%  (p=0.000 n=3+3)
WindowFunctions/(func:cume_dist,_ndv:1000,_rows:100000)-12        3.01k ± 0%    23.83k ± 0%   +692.54%  (p=0.000 n=3+3)
WindowFunctions/(func:percent_rank,_ndv:1000,_rows:100000)-12     3.01k ± 0%    23.85k ± 0%   +693.17%  (p=0.000 n=3+3)
WindowFunctions/(func:ntile,_ndv:1000,_rows:100000)-12            3.01k ± 0%     3.00k ± 0%       ~     (zero variance)
WindowFunctions/(func:lead,_ndv:1000,_rows:100000)-12             3.01k ± 0%   111.01k ± 0%  +3591.58%  (p=0.000 n=3+3)
WindowFunctions/(func:first_value,_ndv:1000,_rows:100000)-12      3.00k ± 0%     3.00k ± 0%       ~     (zero variance)
WindowFunctions/(func:last_value,_ndv:1000,_rows:100000)-12       3.00k ± 0%     3.00k ± 0%       ~     (zero variance)
WindowFunctions/(func:nth_value,_ndv:1000,_rows:100000)-12        3.00k ± 0%     3.00k ± 0%       ~     (zero variance)
```

Related changes

 - Need to cherry-pick to the release branch