@SunRunAway 2019-08-08
The unit test fails because it relies on #11673 
Please ignore it and merge this pr firstly, then merge #11673