@SunRunAway 2019-08-08
/run-integration-common-test tidb-test=pr/872