@SunRunAway 2019-08-08
/run-all-tests tidb-test=pr/872