@SunRunAway 2019-08-13
/run-all-tests tidb-test=pr/872